MyGUI-3.2.0.ebuild: fix installation of library files
cmake-utils.eclass set CMAKE_BUILD_TYPE to the "Gentoo" value
if the variable is not set. The MyGUI cmake script takes
this value and has some logic how to NOT install the libraries
if the values are NOT "ReleaseNone", etc.

After some headaches and a night crash course in cmake, ebuilds
and the rest, fix the problem by just setting CMAKE_BUILD_TYPE
to a void value in the ebuild.